+++ suripu-core/src/main/java/com/hello/suripu/core/db/util/Bucketing.java
@@ -80,9 +80,9 @@ public class Bucketing {
 
             float sensorValue = 0;
             if(sensorName.equals("humidity")) {
-                sensorValue = DeviceData.dbIntToFloat(deviceData.ambientHumidity);
+                sensorValue = DataUtils.dbIntToFloat(deviceData.ambientHumidity);
             } else if(sensorName.equals("temperature")) {
-                sensorValue = DeviceData.dbIntToFloat(deviceData.ambientTemperature);
+                sensorValue = DataUtils.calibrateTemperature(deviceData.ambientTemperature);
             } else if (sensorName.equals("particulates")) {
                 sensorValue = (float) DataUtils.convertRawDustCountsToAQI(deviceData.ambientDustMax, deviceData.firmwareVersion);
             } else if (sensorName.equals("light")) {
@@ -132,8 +132,8 @@ public class Bucketing {
 
             final float lightValue = (float) deviceData.ambientLight;
             final float soundValue = DataUtils.dbIntToFloatAudioDecibels(deviceData.audioPeakBackgroundDB);
-            final float humidityValue = DeviceData.dbIntToFloat(deviceData.ambientHumidity);
-            final float temperatureValue = DeviceData.dbIntToFloat(deviceData.ambientTemperature);
+            final float humidityValue = DataUtils.dbIntToFloat(deviceData.ambientHumidity);
+            final float temperatureValue = DataUtils.calibrateTemperature(deviceData.ambientTemperature);
             final float particulatesValue = (float) DataUtils.convertRawDustCountsToAQI(deviceData.ambientDustMax, deviceData.firmwareVersion);
 
             populatedMap.addSample(newKey, deviceData.offsetMillis,
