+++ suripu-core/src/main/java/com/hello/suripu/core/models/SensorReading.java
@@ -3,9 +3,6 @@ package com.hello.suripu.core.models;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.Objects;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class SensorReading {
 
     @JsonProperty("name")
@@ -22,15 +19,6 @@ public class SensorReading {
         this.name = name;
         this.value = value;
         this.unit = unit;
-    }
-
-    public static List<SensorReading> fromDeviceData(final DeviceData deviceData) {
-        final List<SensorReading> readings = new ArrayList<>();
-        readings.add(new SensorReading("temperature", Math.round(DeviceData.dbIntToFloat(deviceData.ambientTemperature)), CurrentRoomState.State.Unit.CELCIUS));
-        readings.add(new SensorReading("particulates", Math.round(DeviceData.dbIntToFloat(deviceData.ambientAirQuality)), CurrentRoomState.State.Unit.PPM));
-        readings.add(new SensorReading("humidity", Math.round(DeviceData.dbIntToFloat(deviceData.ambientHumidity)), CurrentRoomState.State.Unit.PERCENT));
-
-        return readings;
     }
 
     @Override
