+++ suripu-core/src/main/java/com/hello/suripu/core/util/DataUtils.java
@@ -11,6 +11,8 @@ public class DataUtils{
     private static final float MAX_DUST_ANALOG_VALUE = 4096;
     public static final float DUST_FLOAT_TO_INT_MULTIPLIER = 1000000f;
     public static final float AUDIO_FLOAT_TO_INT_MULTIPLIER = 1000.0f; // 3 decimal places
+    public static final float FLOAT_2_INT_MULTIPLIER = 100;
+    private static final float TEMPERATURE_CALIBRATION_FACTOR_IN_CELSIUS = 278;
 
     // AQI ranges from 0 to 500;
     // see http://www.sparetheair.com/publications/AQI_Lookup_Table-PM25.pdf
@@ -104,4 +106,15 @@ public class DataUtils{
 
     public static float dbIntToFloatAudioDecibels(final int valueFromDatabase) { return ((float) valueFromDatabase) / AUDIO_FLOAT_TO_INT_MULTIPLIER; }
 
+    public static float calibrateTemperature(final int valueFromDatabase) {
+        return valueFromDatabase - TEMPERATURE_CALIBRATION_FACTOR_IN_CELSIUS / FLOAT_2_INT_MULTIPLIER;
+    }
+
+    public static int floatToDBInt(final float value){
+        return (int)(value * FLOAT_2_INT_MULTIPLIER);
+    }
+
+    public static float dbIntToFloat(final int valueFromDB){
+        return valueFromDB / FLOAT_2_INT_MULTIPLIER;
+    }
 }
